Better (?) temporary role selector
Also, added a user display component (the one on the top-right) to allow easier project-wide changes to its styling. It also has the updated role selector. See example in "pages/Faculty/FacultyChangeGrades.tsx".

For now, the new temporary role selector stores the role in the 'localStorage', this should be noted in the future upon working with the backend team as this is insecure.